fix: location field should use forward slashed (in windows too)
When writing the ‘location’ field in the index.json file, it depends on
the operating system. This causes backslashes to appear in windows
instead of the forward slashes that should be used according to the
schema specification. This fix forces the use of the appropriate forward
slashes regardless of the operating system on which the crawler is
running.

Signed-off-by: dsantose <david.santos-esteban@capgemini.com>