feat: lazy instantiation of crawler
The vague instantiation of the crawler allows that when the source where
the vex files are located (URL field in the crawler) is indicated
directly, there is no need to instantiate the crawler. This allows
collaterally to indicate to the crawler packages that are not currently
supported (but do exist in the purl specification) and to generate the
necessary data for the hub.

Signed-off-by: dsantose <david.santos-esteban@capgemini.com>